requirements.txt
numpy==1.10.4
PyAudio==0.2.9
scipy==0.17.0
rpi_ws281x==1.0.3
Adafruit-ADS1x15==1.0.2